Rename web-api file
