+++ bot/classes/guild.js
@@ -1,7 +1,7 @@
 const GuildModel = require("./../../models/guild.js");
 const UserRole = require("./user-role.js");
 const Permission = require("./permission.js");
-const webAPI = require("./../apis/web-api.js").getWebAPI("discotron-dashboard");
+const webAPI = require("../../dashboard/backend/api.js").getWebAPI("discotron-dashboard");
 const Plugin = require("./plugin.js");
 const db = require("./../apis/database-crud.js");
 const Logger = require("../utils/logger.js");
