+++ dashboard/config.json
@@ -1,0 +1,5 @@
+{
+    "webServer": {
+        "port": 47131
+    }
+}
